requirements.txt
playwright==1.46.0
pydantic==2.8.2
python-telegram-bot==21.4
python-dotenv==1.0.1
pdfplumber==0.11.4
rapidfuzz==3.9.6
httpx==0.27.0
google-generativeai==0.7.2
uvloop==0.19.0; platform_system != "Windows"
pydantic==2.8.2